Add files via upload
PyTorch is an open source machine learning library for Python and is completely based on Torch. It is primarily used for applications such as natural language processing. PyTorch is developed by Facebook's artificial-intelligence research group along with Uber's "Pyro" software for the concept of in-built probabilistic programming.

Audience
This tutorial has been prepared for python developers who focus on research and development with machinelearning algorithms along with natural language processing system. The aim of this tutorial is to completely describe all concepts of PyTorch and realworld examples of the same.

Prerequisites
Before proceeding with this tutorial, you need knowledge of Python and Anaconda framework (commands used in Anaconda). Having knowledge of artificial intelligence concepts will be an added advantage.